clean openssl test

+++ picoquictest/openssl_test.c
@@ -74,7 +74,7 @@ static int openssl_cert_test_one(char const * cert_file_name, int expected_count
     if (ret == 0) {
         ptls_iovec_t* certs_vec = picoquic_openssl_get_certs_from_file(cert_file_path, &count);
 
-        if (count != expected_count) {
+        if ((int)count != expected_count) {
             DBG_PRINTF("Expected %d certs for %s, got=%d", expected_count, cert_file_path, count);
             ret = -1;
         }
